OO-2614: delete the file of an assessmentItem only if it isn't referenced any where

+++ src/main/java/org/olat/ims/qti21/ui/editor/AssessmentTestComposerController.java
@@ -1231,8 +1231,11 @@ public class AssessmentTestComposerController extends MainLayoutBasicController 
 			RootNodeLookup<AssessmentItem> rootNode = resolvedAssessmentItem.getItemLookup();
 			if(rootNode != null) {
 				URI itemUri = rootNode.getSystemId();
-				File itemFile = new File(itemUri);
-				deleted = itemFile.delete();
+				List<AssessmentItemRef> itemRefs = resolvedAssessmentTest.getItemRefsBySystemIdMap().get(itemUri);
+				if(itemRefs.size() <= 1) {
+					File itemFile = new File(itemUri);
+					deleted = itemFile.delete();
+				}
 			}
 		}
 		if(deleted) {
